requirements.txt
aiohttp
python-dotenv
requests
python-telegram-bot
